requirements.txt
pillow
requests
tk
tcl
customtkinter